Add NotFound custom exception

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,5 @@
-using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -8,7 +8,7 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
 {
   private readonly ILeaveTypeRepository _leaveTypeRepository;
 
-  public DeleteLeaveTypeCommandHandler( ILeaveTypeRepository leaveTypeRepository)
+  public DeleteLeaveTypeCommandHandler(ILeaveTypeRepository leaveTypeRepository)
   {
     _leaveTypeRepository = leaveTypeRepository;
   }
@@ -17,12 +17,14 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
   {
     // Retrieve domain entity object
     var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
-    
+
     // Verify that the entity exists
-    
+    if (leaveTypeToDelete == null)
+      throw new NotFoundException(nameof(LeaveType), request.Id);
+
     // Delete from database
     await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
-    
+
     return Unit.Value;
   }
 }
